# swagger.yaml
openapi: 3.0.3
info:
  title: GAS API - Goals, Actions, Steps
  description: |
    Production-ready REST API for the GAS task management and scheduling application.

    ## Features
    - JWT-based authentication
    - Kanban board task management
    - Meeting scheduling and availability
    - Admin panel for user and system management

    ## Authentication
    All endpoints except `/auth/login` and `/auth/register` require a valid JWT token.
    Include the token in the Authorization header: `Bearer <token>`

  version: 1.0.0
  contact:
    name: API Support
    email: support@gas-app.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.gas-app.com
    description: Production server
...
paths:
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve authenticated user information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # TASKS ENDPOINTS
  # ==========================================
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks organized by columns
      description: Retrieve all tasks grouped by Kanban columns (Goals, Actions, Steps, Completed)
      operationId: getTasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Column'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Add a new task to a specific column
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
            examples:
              goal:
                summary: Create a goal
                value:
                  title: Increase sales by 20%
                  description: Focus on enterprise clients
                  priority: high
                  assignee: John Doe
                  columnId: goals
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{taskId}:
    patch:
      tags:
        - Tasks
      summary: Update a task
      description: Update task properties
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Permanently delete a task
      operationId: deleteTask
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{taskId}/move:
    patch:
      tags:
        - Tasks
      summary: Move task to different column
      description: Move a task from one Kanban column to another
      operationId: moveTask
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - toColumnId
              properties:
                toColumnId:
                  type: string
                  enum: [goals, actions, steps, completed]
                  description: Target column ID
                  example: completed
      responses:
        '200':
          description: Task moved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{taskId}/toggle-complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion status
      description: Mark task as completed or incomplete
      operationId: toggleTaskComplete
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Task ID
      responses:
        '200':
          description: Task status toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # MEETINGS ENDPOINTS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT-based authentication. Include the token in the Authorization header:
        `Authorization: Bearer <token>`
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "1"
        name:
          type: string
          description: User full name
          example: John Doe
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@gas-app.com
        role:
          type: string
          enum: [admin, operator, viewer]
          description: User role
          example: admin
        status:
          type: string
          enum: [active, inactive]
          description: User account status
          example: active
        lastActive:
          type: string
          format: date-time
          description: Last activity timestamp
          example: "2026-05-13T10:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-01-01T00:00:00Z"
...
    Task:
      type: object
      properties:
        id:
          type: string
          description: Unique task identifier
          example: "1"
        title:
          type: string
          description: Task title
          example: Increase sales by 20%
        description:
          type: string
          description: Task description
          example: Focus on enterprise clients
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
          example: high
        assignee:
          type: string
          description: Assigned user name
          example: John Doe
        columnId:
          type: string
          enum: [goals, actions, steps, completed]
          description: Current Kanban column
          example: goals
        date:
          type: string
          format: date
          description: Task due date
          example: "2026-05-15"
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Task time (HH:mm format)
          example: "10:00"
        type:
          type: string
          enum: [goal, action, step]
          description: Task type for calendar
          example: goal
        completed:
          type: boolean
          description: Completion status
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-05-13T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-13T10:00:00Z"
    Column:
      type: object
      properties:
        id:
          type: string
          enum: [goals, actions, steps, completed]
          description: Column identifier
          example: goals
        title:
          type: string
          description: Column display title
          example: Goals
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: Tasks in this column
        order:
          type: integer
          description: Display order
          example: 0
    TaskCreateRequest:
      type: object
      required:
        - title
        - priority
        - assignee
        - columnId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Complete project documentation
        description:
          type: string
          maxLength: 1000
          example: Write comprehensive API documentation
        priority:
          type: string
          enum: [low, medium, high]
          example: medium
        assignee:
          type: string
          example: John Doe
        columnId:
          type: string
          enum: [goals, actions, steps, completed]
          example: actions
        date:
          type: string
          format: date
          example: "2026-05-15"
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "14:00"
        type:
          type: string
          enum: [goal, action, step]
          example: action
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        completed:
          type: boolean

    # ==========================================
    # MEETING SCHEMAS
    # ==========================================
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error message
          example: Validation failed
        code:
          type: string
          description: Error code for programmatic handling
          example: VALIDATION_ERROR
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
          example:
            email:
              - Email is required
              - Email must be valid
            password:
              - Password must be at least 6 characters
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed
            code: VALIDATION_ERROR
            errors:
              title:
                - Title is required
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication required
            code: UNAUTHORIZED
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Resource not found
            code: NOT_FOUND
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: An unexpected error occurred
            code: INTERNAL_SERVER_ERROR
security:
  - BearerAuth: []